fix: remove LikeControllerV2 for merging to ArchiveLikeControllerV2
